+++ src/routes/rentals/[slug].svelte
@@ -1,8 +1,9 @@
 <script type="ts" context="module">
 	import type { Rental } from 'src/models/rental.model';
 	import Jumbo from '../../components/jumbo.svelte';
-	import RentalImage from '../../components/rental/image.svelte';
 	import Map from '../../components/map.svelte';
+	import RentalImage from '../../components/rental/image.svelte';
+	import ShareButton from '../../components/share-button.svelte';
 
 	const COMMUNITY_CATEGORIES = ['Condo', 'Townhouse', 'Apartment'];
 
@@ -34,14 +35,13 @@
 <Jumbo>
 	<h2>{rental.title}</h2>
 	<p>Nice find! This looks like a nice place to stay near {rental.city}.</p>
-	<a
-		href="https://twitter.com"
-		target="_blank"
-		rel="external nofollow noopener noreferrer"
-		class="share button"
+	<ShareButton
+		text="Check out {rental.title} on Super Rentals!"
+		hashtags="vacation,travel,authentic,blessed,superrentals"
+		via="sveltejs"
 	>
 		Share on Twitter
-	</a>
+	</ShareButton>
 </Jumbo>
 
 <article class="rental detailed">